// src/models/property.ts
export class Property {
    id: number;
    title: string;
    userID: number;
    username: string;
    phone: string;
    type: string;
    province: string;
    city: string;
    address: string;
    meters: number;
    price: number;
    date: string;
    deal: string;
    description: string;
}